Safari-611.1.10/LayoutTests/fast/dom/HTMLElement/attr-missing-parent-ancestor-missing.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id="result"></div>
<div id="div">
  <div>
    <div>
      <p id="p"></p>
    </div>
  </div>
</div>
<script>
description('When contentEditable is missing, it maps to "inherit" state and the element is NOT editable if its parent/ancestor\'s contentEditable is also missing.')

shouldBe('document.getElementById("p").hasAttribute("contentEditable")','false');
shouldBe('document.getElementById("p").contentEditable', '"inherit"');
shouldBe('document.getElementById("p").isContentEditable', 'false');
shouldBe('window.getComputedStyle(p, "").getPropertyValue("-webkit-user-modify")', '"read-only"');

document.getElementById("div").style.display= 'none';
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
